refactor: move `parse_params` to `rpc_router.go`
