+++ clearnode/rpc_router.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"encoding/json"
+	"fmt"
 	"math/big"
 	"time"
 
@@ -254,4 +255,18 @@ func (r *RPCRouter) HandleGetRPCHistory(c *RPCContext) {
 
 	c.Succeed(c.Message.Req.Method, response)
 	logger.Info("RPC history retrieved", "userID", c.UserID, "entryCount", len(response))
+}
+
+func parseParams(params []any, unmarshalTo any) error {
+	if len(params) > 0 {
+		paramsJSON, err := json.Marshal(params[0])
+		if err != nil {
+			return fmt.Errorf("failed to parse parameters: %w", err)
+		}
+		err = json.Unmarshal(paramsJSON, &unmarshalTo)
+		if err != nil {
+			return err
+		}
+	}
+	return validate.Struct(unmarshalTo)
 }
